requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
Django==4.2.4
djangorestframework==3.14.0
drf-yasg==1.21.7
gTTS==2.3.2
idna==3.4
inflection==0.5.1
packaging==23.1
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
